### 2024-08-05 #86 @ValterH
[Re] Learning with Noisy Labels Revisited: A Study Using Real-World Human Annotations
**Original article:** [Learning with Noisy Labels Revisited: A Study Using Real-World Human Annotations](https://openreview.net/pdf?id=TBWA6PLJZQm)

**Openreview:** https://openreview.net/forum?id=GKZ2leags0

**Scientific domain:**  Machine Learning
**Suggested editor:**  @rougier
