### 2024-10-11 #86 @rougier
@ValterH Very sorry for such long delay, for some reasons, I did not see your submission.
Is your paper part of the ML Reproducibility Challenge ? If yes, do you have open peer review already?
