### 2025-02-07 a4e89e3
add optional reproducibility requirements
