### 2025-02-10 71b4ff1
add missing configs
