### 2025-02-10 fcd7412
refactor PES utils
