### 2025-02-10 b3dd07b
Add missing paper configs
