+++ src/reproducibility/noisy_labels/configs/cifar100n/__init__.py
@@ -10,3 +10,18 @@ from .PES_semi import PES_semi_config
 from .SOP import SOP_config
 from .SOP_plus import SOP_plus_config
 from .volminnet import volminnet_config
+
+__all__ = [
+    "CAL_config",
+    "CE_config",
+    "co_teaching_config",
+    "co_teaching_plus_config",
+    "divide_mix_config",
+    "divide_mix_clean_config",
+    "ELR_config",
+    "ELR_plus_config",
+    "PES_semi_config",
+    "SOP_config",
+    "SOP_plus_config",
+    "volminnet_config",
+]
