+++ src/reproducibility/noisy_labels/configs/benchmark_configs/cifar10n/CE.py
@@ -1,0 +1,27 @@
+from noisypy.methods.learning_strategies.CE.CE import CE
+from torch.optim import SGD
+from torch.optim.lr_scheduler import MultiStepLR
+
+from ..base import BenchmarkConfigCIFAR10N
+
+
+class CE_config(BenchmarkConfigCIFAR10N):
+
+    learning_strategy_cls = CE
+
+    optimizer_cls=SGD
+    optimizer_args=dict(
+        lr=0.1,
+        momentum=0.9,
+        weight_decay=5e-4,
+    )
+    scheduler_cls=MultiStepLR
+    scheduler_args=dict(
+        milestones=[60],
+        gamma=0.1
+    )
+
+    trainer_args = {
+        **BenchmarkConfigCIFAR10N.trainer_args,
+        "max_epochs": 100,
+    }

[…]
