### 2025-02-10 e3c3fef
update reproducibility requirements
