### 2025-02-10 d7f786d
Scope: remove EuroSAT.
