### 2025-02-10 443afd6
Remove old datamodule tests.
