+++ src/noisypy/data/datamodule.py
@@ -31,7 +31,7 @@ class MultiSampleDataModule(LightningDataModule):
 
     @property
     def num_train_samples(self) -> int:
-        return len(self.train_datasets[0]) # TODO: bug if different lengths - will this ever happen?
+        return len(self.train_datasets[0])
     
     @property
     def num_val_samples(self) -> int:
@@ -46,7 +46,6 @@ class MultiSampleDataModule(LightningDataModule):
         return self.train_datasets[0].num_classes
 
     def setup(self, stage: str = None) -> None:
-        # TODO: figure out how to handle setup
         pass
     
     def train_dataloader(self) -> TRAIN_DATALOADERS:
