+++ tests/test_units/test_data/test_pipelines/test_noise/test_noises.py
@@ -2,7 +2,6 @@ import torch
 
 from noisypy.data.pipelines.noise.noises import InstanceNoise, AsymmetricNoise, SymmetricNoise, LambdaNoise
 
-# TODO remember mat2 and calculate 
 # teoretical guarantees for n_samples to converge
 
 def test_persistance():
