+++ src/reproducibility/learning_strategies/divide_mix/divide_mix_cifar10_noise.py
@@ -31,7 +31,6 @@ class divide_mix_cifar10_noise(divide_mix_cifar10_clean):
         max_epochs=300,
         reload_dataloaders_every_n_epochs=1,
         deterministic=True,
-        # TODO: move to logs when we are not running from src/
         logger=CSVLogger("../logs", name="divide_mix_symmetric")
     )
 
