### 2025-02-10 aba5ee0
add val loss logging to CoTeaching
