### 2025-02-10 100c82e
Fix type[] -> Type[] in src.
