+++ src/noisypy/methods/learning_strategies/jocor/utils.py
@@ -1,6 +1,4 @@
 import torch
-from torch import nn
-from torch.nn import functional as F
 import numpy as np
 import torch.nn.functional as F
 
