+++ src/noisypy/methods/learning_strategies/peer_loss/peer_loss.py
@@ -1,4 +1,4 @@
-from typing import Any
+from typing import Any, Type
 from copy import deepcopy
 from functools import partial
 
@@ -18,8 +18,8 @@ class PeerLoss(MultiStageLearningStrategyModule):
 
     def __init__(self, datamodule: L.LightningDataModule,
                  classifier_cls: type, classifier_args: dict,
-                 optimizer_cls: list[type[Optimizer]], optimizer_args: list[dict],
-                 scheduler_cls: list[type[LRScheduler]], scheduler_args: list[dict],
+                 optimizer_cls: list[Type[Optimizer]], optimizer_args: list[dict],
+                 scheduler_cls: list[Type[LRScheduler]], scheduler_args: list[dict],
                  stage_epochs: list[int], warmup_criterion: nn.Module = CrossEntropyLossStable,
                  *args: Any, **kwargs: Any) -> None:
         super().__init__(
