+++ src/main.py
@@ -8,8 +8,8 @@ def run(method_config):
 
 if __name__ == '__main__':
     parser = argparse.ArgumentParser()
-    parser.add_argument('--config', type=str, help='Path to the config .py file.', default="configs/methods/cores2/cores_cifar10_clean.py", required=False)
-    parser.add_argument('--method', type=str, help='Name of MethodConfig class to run.', default="cores_cifar10_clean", required=False)
+    parser.add_argument('--config', type=str, help='Path to the config .py file.', default="noisypy/configs/methods/cores2/cores2_cifar10_clean.py", required=False)
+    parser.add_argument('--method', type=str, help='Name of MethodConfig class to run.', default="cores2_cifar10_clean_config", required=False)
     args = parser.parse_args()
 
     modulename = SourceFileLoader("cfg", str(args.config)).load_module()
