+++ src/reproducibility/learning_strategies/FBT/forwardT_cifar10_clean.py
@@ -1,4 +1,4 @@
-from aim.pytorch_lightning import AimLogger
+from lightning.pytorch.loggers import CSVLogger
 from torch.optim import SGD
 from torch.optim.lr_scheduler import MultiStepLR
 
@@ -34,7 +34,7 @@ class forwardT_cifar10_clean(MethodConfig):
         max_epochs=240,
         deterministic=True,
         num_sanity_val_steps=0,
-        logger=AimLogger(experiment="forwardT")
+        logger=CSVLogger("../logs", name="forwardT_clean"),
     )
 
     seed = 1337
