+++ src/reproducibility/learning_strategies/co_teaching/co_teaching_plus_reprod.py
@@ -1,6 +1,5 @@
-from aim.pytorch_lightning import AimLogger
+from pytorch_lightning.loggers import CSVLogger
 
-import torch
 from torch.optim import Adam
 from torch.optim.lr_scheduler import LambdaLR
 from torchvision.transforms import ToTensor
@@ -71,7 +70,7 @@ class co_teaching_plus_reprod(MethodConfig):
     trainer_args = dict(
         max_epochs=200,
         deterministic=True,
-        logger=AimLogger(experiment="CoTeachingPlus")
+        logger=CSVLogger("../logs", name="CoTeachingPlus")
     )
 
     seed = 1
