+++ src/reproducibility/memorization_effects/memo.py
@@ -1,26 +1,30 @@
-# %%
+import argparse
+
 import torch
 from torch import nn
 from torch.nn import functional as F
 from torchvision.datasets import CIFAR10
-# from torchvision.models.resnet import resnet34
-from torchvision.models.feature_extraction import create_feature_extractor, get_graph_node_names
 from torchvision import transforms
-
-import numpy as np
 from tqdm import tqdm
 
 from utils import ResNet34 as resnet34
 
 device = "cuda" if torch.cuda.is_available() else "cpu"
 
-# %%
-noisy_path = "/d/hpc/projects/FRI/vh0153/noisy-labels/src/reproducibility/noisy_labels/data/noisylabels/CIFAR-10_human.pt"
-noisy_labels = {k: torch.tensor(v) for k, v in torch.load(noisy_path).items()}
-keys_to_test = ["aggre_label", "random_label1", "worse_label"]
-noisy_labels
+argparser = argparse.ArgumentParser()
+argparser.add_argument("noise_label", type=str, help="noise label type", choices=["aggre_label", "random_label1", "worse_label"])
+argparser.add_argument("--seed", type=int, default=1, help="random seed")
+argparser.add_argument("--device", type=int, default=0, help="device")
 
-# %%
+args = argparser.parse_args()
+seed = args.seed
+noise_label = args.noise_label
+device = f"cuda:{args.device}" if torch.cuda.is_available() else "cpu"
+
+noisy_path = "../../../data/noisylabels/CIFAR-10_human.pt"
+noisy_labels = {k: torch.tensor(v) for k, v in torch.load(noisy_path).items()}
+
+
 cifar10_train_transform = transforms.Compose([
     transforms.RandomCrop(32, padding=4),
     transforms.RandomHorizontalFlip(),
@@ -28,11 +32,11 @@ cifar10_train_transform = transforms.Compose([
     transforms.Normalize((0.4914, 0.4822, 0.4465), (0.2023, 0.1994, 0.2010)),
 ])
 
-clean_dataset = CIFAR10("/d/hpc/projects/FRI/vh0153/noisy-labels/src/reproducibility/noisy_labels/data/cifar", train=True, transform=cifar10_train_transform)
+clean_dataset = CIFAR10("../../../data/cifar", train=True, transform=cifar10_train_transform)
 clean_labels = torch.tensor(clean_dataset.targets)
 clean_labels
 
-# %%
+
 def generate_synthetic_labels(clean_labels, noisy_labels, seed=1337):
     # generate transition matrix T
     num_classes = 10
@@ -48,7 +52,7 @@ def generate_synthetic_labels(clean_labels, noisy_labels, seed=1337):
     synthetic_labels = torch.multinomial(T[clean_labels], 1, generator=gen).squeeze()
     return synthetic_labels
 
-# %%
+
 def train_and_record_memorization(clean_labels, noisy_labels, num_epochs):
     dataset = clean_dataset
     dataset.targets = noisy_labels
@@ -59,17 +63,13 @@ def train_and_record_memorization(clean_labels, noisy_labels, num_epochs):
     # use the same optimizer as in the CE config
     optimizer = torch.optim.SGD(
         model.parameters(), 
-        lr=0.1,
-        momentum=0.9,
-        weight_decay=5e-4)
-    # use the same scheduler as in the CE config
-    scheduler = torch.optim.lr_scheduler.MultiStepLR(
-        optimizer, 
-        milestones=[60], 
-        gamma=0.1)
+        lr=1,
+        weight_decay=1e-4)
+    lambda_lr = lambda epoch:  0.1 ** (epoch / 50)
[…]
