+++ src/reproducibility/logs/benchmark/cifar100n_clean_benchmark_config/divide_mix_clean_config_42/version_0/hparams.yaml
@@ -1,0 +1,7 @@
+alpha: 4
+lambda_u: 0
+noise_rate: 0.5
+noise_type: asymmetric
+p_thresh: 0.5
+temperature: 0.5
+warmup_epochs: 30
