+++ src/reproducibility/logs/benchmark/cifar100n_noisy_benchmark_config/SOP_config_1/version_0/hparams.yaml
@@ -1,0 +1,8 @@
+lr_u: 1
+lr_v: 10
+overparam_mean: 0.0
+overparam_momentum: 0
+overparam_std: 1.0e-08
+overparam_weight_decay: 0
+ratio_balance: 0
+ratio_consistency: 0
