+++ src/reproducibility/logs/benchmark/cifar100n_clean_benchmark_config/PES_semi_config_1337/version_0/hparams.yaml
@@ -1,0 +1,7 @@
+PES_lr: 0.0001
+T2: 5
+alpha: 4
+lambda_u: 75
+model_type: pytorch_resnet
+temperature: 0.5
+warmup_epochs: 35
