+++ src/reproducibility/logs/baseline/rand1_1337.txt
@@ -1,0 +1,1010 @@
+Files already downloaded and verified
+Loaded random_label1 from ./data/CIFAR-10_human.pt.
+The overall noise rate is 0.17234000000000005
+noisy labels loaded from ./data/CIFAR-10_human.pt
+The noisy data ratio in each class is [0.10042 0.11116 0.10292 0.09388 0.0889  0.10766 0.0935  0.1051  0.09842
+ 0.09804]
+over all noise rate is  0.17234
+train_labels: 50000 [6, 9, 9, 4, 1, 1, 2, 7, 8, 3]
+building model...
+building model done
+epoch 0
+Epoch [1/100], Iter [50/390] Training Accuracy: 17.1875, Loss: 2.2110
+Epoch [1/100], Iter [100/390] Training Accuracy: 18.7500, Loss: 2.0945
+Epoch [1/100], Iter [150/390] Training Accuracy: 19.5312, Loss: 2.0337
+Epoch [1/100], Iter [200/390] Training Accuracy: 23.4375, Loss: 2.0222
+Epoch [1/100], Iter [250/390] Training Accuracy: 26.5625, Loss: 1.9739
+Epoch [1/100], Iter [300/390] Training Accuracy: 26.5625, Loss: 1.9876
+Epoch [1/100], Iter [350/390] Training Accuracy: 26.5625, Loss: 2.0413
+train acc on train images is  22.117966751918157
+test acc on test images is  32.92
+epoch 1
+Epoch [2/100], Iter [50/390] Training Accuracy: 26.5625, Loss: 1.8823
+Epoch [2/100], Iter [100/390] Training Accuracy: 31.2500, Loss: 1.7367
+Epoch [2/100], Iter [150/390] Training Accuracy: 30.4688, Loss: 1.7899
+Epoch [2/100], Iter [200/390] Training Accuracy: 39.0625, Loss: 1.6524
+Epoch [2/100], Iter [250/390] Training Accuracy: 32.8125, Loss: 1.7635
+Epoch [2/100], Iter [300/390] Training Accuracy: 35.1562, Loss: 1.7503
+Epoch [2/100], Iter [350/390] Training Accuracy: 40.6250, Loss: 1.7108
+train acc on train images is  33.319213554987215
+test acc on test images is  40.58
+epoch 2
+Epoch [3/100], Iter [50/390] Training Accuracy: 35.1562, Loss: 1.7594
+Epoch [3/100], Iter [100/390] Training Accuracy: 39.0625, Loss: 1.6480
+Epoch [3/100], Iter [150/390] Training Accuracy: 44.5312, Loss: 1.5953
+Epoch [3/100], Iter [200/390] Training Accuracy: 42.1875, Loss: 1.6600
+Epoch [3/100], Iter [250/390] Training Accuracy: 42.9688, Loss: 1.6717
+Epoch [3/100], Iter [300/390] Training Accuracy: 37.5000, Loss: 1.6128
+Epoch [3/100], Iter [350/390] Training Accuracy: 42.1875, Loss: 1.6683
+train acc on train images is  40.121882992327365
+test acc on test images is  44.02
+epoch 3
+Epoch [4/100], Iter [50/390] Training Accuracy: 49.2188, Loss: 1.6350
+Epoch [4/100], Iter [100/390] Training Accuracy: 46.0938, Loss: 1.5941
+Epoch [4/100], Iter [150/390] Training Accuracy: 40.6250, Loss: 1.6902
+Epoch [4/100], Iter [200/390] Training Accuracy: 49.2188, Loss: 1.5469
+Epoch [4/100], Iter [250/390] Training Accuracy: 45.3125, Loss: 1.6198
[…]
