+++ src/reproducibility/logs/baseline/rand2_1337.txt
@@ -1,0 +1,1010 @@
+Files already downloaded and verified
+Loaded random_label2 from ./data/CIFAR-10_human.pt.
+The overall noise rate is 0.18122000000000005
+noisy labels loaded from ./data/CIFAR-10_human.pt
+The noisy data ratio in each class is [0.10148 0.10994 0.10398 0.09388 0.08958 0.1071  0.09398 0.10366 0.0983
+ 0.0981 ]
+over all noise rate is  0.18122
+train_labels: 50000 [6, 9, 9, 4, 1, 1, 2, 7, 8, 3]
+building model...
+building model done
+epoch 0
+Epoch [1/100], Iter [50/390] Training Accuracy: 10.9375, Loss: 2.2951
+Epoch [1/100], Iter [100/390] Training Accuracy: 9.3750, Loss: 2.3153
+Epoch [1/100], Iter [150/390] Training Accuracy: 25.7812, Loss: 2.0233
+Epoch [1/100], Iter [200/390] Training Accuracy: 17.9688, Loss: 2.1280
+Epoch [1/100], Iter [250/390] Training Accuracy: 28.9062, Loss: 1.9691
+Epoch [1/100], Iter [300/390] Training Accuracy: 24.2188, Loss: 2.1171
+Epoch [1/100], Iter [350/390] Training Accuracy: 23.4375, Loss: 2.1063
+train acc on train images is  19.366608056265985
+test acc on test images is  31.39
+epoch 1
+Epoch [2/100], Iter [50/390] Training Accuracy: 28.9062, Loss: 1.9462
+Epoch [2/100], Iter [100/390] Training Accuracy: 34.3750, Loss: 1.8533
+Epoch [2/100], Iter [150/390] Training Accuracy: 25.7812, Loss: 1.9292
+Epoch [2/100], Iter [200/390] Training Accuracy: 45.3125, Loss: 1.6419
+Epoch [2/100], Iter [250/390] Training Accuracy: 34.3750, Loss: 1.8534
+Epoch [2/100], Iter [300/390] Training Accuracy: 37.5000, Loss: 1.7562
+Epoch [2/100], Iter [350/390] Training Accuracy: 35.9375, Loss: 1.7690
+train acc on train images is  32.23785166240409
+test acc on test images is  41.62
+epoch 2
+Epoch [3/100], Iter [50/390] Training Accuracy: 36.7188, Loss: 1.5891
+Epoch [3/100], Iter [100/390] Training Accuracy: 35.1562, Loss: 1.7025
+Epoch [3/100], Iter [150/390] Training Accuracy: 42.9688, Loss: 1.5886
+Epoch [3/100], Iter [200/390] Training Accuracy: 44.5312, Loss: 1.6089
+Epoch [3/100], Iter [250/390] Training Accuracy: 36.7188, Loss: 1.8561
+Epoch [3/100], Iter [300/390] Training Accuracy: 41.4062, Loss: 1.6931
+Epoch [3/100], Iter [350/390] Training Accuracy: 42.1875, Loss: 1.6656
+train acc on train images is  38.54899296675192
+test acc on test images is  43.99
+epoch 3
+Epoch [4/100], Iter [50/390] Training Accuracy: 46.8750, Loss: 1.5721
+Epoch [4/100], Iter [100/390] Training Accuracy: 40.6250, Loss: 1.7440
+Epoch [4/100], Iter [150/390] Training Accuracy: 46.8750, Loss: 1.5941
+Epoch [4/100], Iter [200/390] Training Accuracy: 42.1875, Loss: 1.6306
+Epoch [4/100], Iter [250/390] Training Accuracy: 33.5938, Loss: 1.7334
[…]

[…]
