+++ src/reproducibility/learning_strategies/PESsemi/pes_semi_cifar10_noise.py
@@ -23,7 +23,6 @@ class pes_semi_cifar10n_noise(pes_semi_cifar10n_clean):
         reload_dataloaders_every_n_epochs=1,
         max_epochs=300,
         deterministic=True,
-        # TODO: move to logs when we are not running from src/
         logger=CSVLogger("../logs", name="pes_semi_noise")
     )
 
