+++ src/noisypy/methods/learning_strategies/FBT/FBT.py
@@ -1,4 +1,4 @@
-from typing import Any
+from typing import Any, Type
 from copy import deepcopy
 
 from lightning.pytorch.utilities.types import STEP_OUTPUT
@@ -13,15 +13,13 @@ from torch.optim.lr_scheduler import LRScheduler
 from .utils import ForwardTLoss, BackwardTLoss, estimate_noise_mtx
 from ..base import LearningStrategyWithWarmupModule
 
-# TODO port this to MultiStageLearningStrategyModule
-
 class ForwardT(LearningStrategyWithWarmupModule):
 
     def __init__(self, 
                  datamodule: L.LightningDataModule,
                  classifier_cls: type, classifier_args: dict,
-                 optimizer_cls: type[Optimizer], optimizer_args: dict,
-                 scheduler_cls: type[LRScheduler], scheduler_args: dict,
+                 optimizer_cls: Type[Optimizer], optimizer_args: dict,
+                 scheduler_cls: Type[LRScheduler], scheduler_args: dict,
                  warmup_epochs: int, filter_outliers: bool,
                  *args: Any, **kwargs: Any) -> None:
         super().__init__(
@@ -130,8 +128,8 @@ class BackwardT(ForwardT):
     def __init__(self, 
                  datamodule: L.LightningDataModule, 
                  classifier_cls: type, classifier_args: dict, 
-                 optimizer_cls: type[Optimizer], optimizer_args: dict, 
-                 scheduler_cls: type[LRScheduler], scheduler_args: dict, 
+                 optimizer_cls: Type[Optimizer], optimizer_args: dict, 
+                 scheduler_cls: Type[LRScheduler], scheduler_args: dict, 
                  warmup_epochs: int, filter_outliers: bool, 
                  *args: Any, **kwargs: Any) -> None:
         super().__init__(
