+++ src/reproducibility/noisy_labels/configs/base/wrappers.py
@@ -1,28 +1,29 @@
+from noisypy.configs.base.data import DataConfig
 from noisypy.data.pipelines import AddIndex, DivideMixify, Identity, ShuffleImages, Compose, DoubleAugmentation
-from noisypy.data.datasets.cifar10 import cifar10_train_transform
+from noisypy.data.datasets.cifar10 import cifar10_train_transform, cifar10_test_transform
 from noisypy.methods.learning_strategies.SOPplus.utils import autoaug_paper_cifar10
 
 
-def add_index_wrapper(data_config):
+def add_index_wrapper(data_config: DataConfig) -> DataConfig:
     class new_data_config(data_config):
         dataset_train_augmentation = AddIndex()
 
     return new_data_config
 
-def dividemixify_wrapper(data_config):
+def dividemixify_wrapper(data_config: DataConfig) -> DataConfig:
     class new_data_config(data_config):
         dataset_train_args = [
             dict(mode="all",        transform=cifar10_train_transform),
             dict(mode="all",        transform=cifar10_train_transform),
             dict(mode="unlabeled",  transform=cifar10_train_transform),
             dict(mode="unlabeled",  transform=cifar10_train_transform),
-            dict(mode="all",        transform=cifar10_train_transform),
+            dict(mode="all",        transform=cifar10_test_transform),
         ]
         dataset_train_augmentation = DivideMixify()
 
     return new_data_config
 
-def peer_wrapper(data_config):
+def peer_wrapper(data_config: DataConfig) -> DataConfig:
     class new_data_config(data_config):
         dataset_train_augmentation = [
             Identity(),
@@ -31,7 +32,7 @@ def peer_wrapper(data_config):
     
     return new_data_config
 
-def double_aug_wrapper(data_config):
+def double_aug_wrapper(data_config: DataConfig) -> DataConfig:
     class new_data_config(data_config):
         dataset_train_args = {**data_config.dataset_train_args, "transform": None}
         dataset_train_augmentation = Compose(
