+++ src/noisypy/methods/learning_strategies/pro_mix/pro_mix.py
@@ -1,0 +1,443 @@
+from typing import Any, Type
+import math
+
+from tqdm import tqdm
+import lightning as L
+from lightning.pytorch.utilities.types import STEP_OUTPUT
+import torchmetrics
+import torch
+from torch import Tensor
+from torch.nn import Module
+from torch.nn.functional import one_hot
+from torch.optim import Optimizer
+from torch.optim.lr_scheduler import LRScheduler
+from torch.utils.data import DataLoader
+
+from ..base import LearningStrategyModule
+from .utils import (NegEntropy, linear_rampup, co_divide, debias_output, 
+                    bias_initial, debias_pl, bias_update, BATCH_MAP, end_warmup, 
+                    CE_Soft_Label, ProMixModel)
+from .fmix import FMix
+
+class ProMix(LearningStrategyModule):
+    def __init__(self, datamodule: L.LightningDataModule,
+                 classifier_cls: Type[Module], classifier_args: dict,
+                 optimizer_cls: Type[Optimizer], optimizer_args: dict,
+                 scheduler_cls: Type[LRScheduler], scheduler_args: dict,
+                 warmup_epochs: int, rampup_epochs: int, noise_type: str, 
+                 rho_start: float, rho_end: float, debias_beta_pl: float,
+                 alpha_output: float, tau: float, start_expand: int,
+                 threshold: float, temperature: float,
+                 model_type: str, feat_dim: int,
+                 bias_m: float, *args: Any, **kwargs: Any) -> None:
+        super().__init__(
+            datamodule, classifier_cls, classifier_args,
+            optimizer_cls, optimizer_args, 
+            scheduler_cls, scheduler_args, *args, **kwargs)
+        # saves arguments (hyperparameters) passed to the constructor as self.hparams and logs them to hparams.yaml.
+        self.save_hyperparameters(ignore=["classifier_cls", "classifier_args", "datamodule", 
+                                          "optimizer_cls", "optimizer_args", 
+                                          "scheduler_cls", "scheduler_args"])
+        self.num_training_samples = datamodule.num_train_samples
+        self.num_classes = datamodule.num_classes
+        self.warmup_epochs = warmup_epochs
+        self.rampup_epochs = rampup_epochs
+        self.noise_type = noise_type
+        self.rho_start = rho_start  
+        self.rho_end = rho_end
+        self.debias_beta_pl = debias_beta_pl
+        self.beta = 1
+        self.alpha_output = alpha_output
+        self.tau = tau
+        self.start_expand = start_expand
+        self.threshold = threshold
+        self.temperature = temperature
+        self.bias_m = bias_m
+        self.batch_size = self.datamodule.train_dataloader()[BATCH_MAP['train']].batch_size
+
+        self.fmix = FMix()
+
+        self.pi1 = bias_initial(self.num_classes)
+        self.pi2 = bias_initial(self.num_classes)
+        self.pi1_unrel = bias_initial(self.num_classes)
+        self.pi2_unrel = bias_initial(self.num_classes)
+
+
+        # init model
+        self.model1 = ProMixModel(base_model_cls=classifier_cls, model_type=model_type, feat_dim=feat_dim, **classifier_args)
[…]
