### 2025-02-24 62b3220
Lint and add pre-commit hook
