+++ src/noisypy/methods/learning_strategies/SOP/sop.py
@@ -12,23 +12,52 @@ from .utils import overparametrization_loss
 
 
 class SOP(LearningStrategyModule):
-    def __init__(self, datamodule: L.LightningDataModule,
-                 classifier_cls: type, classifier_args: dict,
-                 optimizer_cls: Type[Optimizer], optimizer_args: dict,
-                 scheduler_cls: Type[LRScheduler], scheduler_args: dict,
-                 ratio_consistency: float, ratio_balance: float, lr_u: float, lr_v: float,
-                 overparam_optimizer_cls: Type[Optimizer], overparam_weight_decay: float, 
-                 overparam_momentum: float, overparam_mean: float, overparam_std: float,
-                 *args: Any, **kwargs: Any) -> None:
+    def __init__(
+        self,
+        datamodule: L.LightningDataModule,
+        classifier_cls: type,
+        classifier_args: dict,
+        optimizer_cls: Type[Optimizer],
+        optimizer_args: dict,
+        scheduler_cls: Type[LRScheduler],
+        scheduler_args: dict,
+        ratio_consistency: float,
+        ratio_balance: float,
+        lr_u: float,
+        lr_v: float,
+        overparam_optimizer_cls: Type[Optimizer],
+        overparam_weight_decay: float,
+        overparam_momentum: float,
+        overparam_mean: float,
+        overparam_std: float,
+        *args: Any,
+        **kwargs: Any,
+    ) -> None:
         super().__init__(
-            datamodule, classifier_cls, classifier_args,
-            optimizer_cls, optimizer_args, 
-            scheduler_cls, scheduler_args, *args, **kwargs)
+            datamodule,
+            classifier_cls,
+            classifier_args,
+            optimizer_cls,
+            optimizer_args,
+            scheduler_cls,
+            scheduler_args,
+            *args,
+            **kwargs,
+        )
         # saves arguments (hyperparameters) passed to the constructor as self.hparams and logs them to hparams.yaml.
         # ratio_consistency, ratio_balance, lr_u, lr_v, overparam_weight_decay, overparam_momentum, overparam_mean, overparam_std
-        self.save_hyperparameters(ignore=["classifier_cls", "classifier_args", "datamodule", 
-                                          "optimizer_cls", "optimizer_args", 
-                                          "scheduler_cls", "scheduler_args", "overparam_optimizer_cls"])
+        self.save_hyperparameters(
+            ignore=[
+                "classifier_cls",
+                "classifier_args",
+                "datamodule",
+                "optimizer_cls",
+                "optimizer_args",
+                "scheduler_cls",
+                "scheduler_args",
+                "overparam_optimizer_cls",
+            ]
+        )
         self.num_training_samples = datamodule.num_train_samples
         self.num_classes = datamodule.num_classes
 
@@ -40,16 +69,27 @@ class SOP(LearningStrategyModule):
         self.overparam_weight_decay = overparam_weight_decay
         self.overparam_optimizer_cls = overparam_optimizer_cls
         self.model = classifier_cls(**classifier_args)
-        self.criterion = overparametrization_loss(num_examp=self.num_training_samples, num_classes=self.num_classes, 
-                                                  ratio_consistency=self.ratio_consistency, ratio_balance=ratio_balance,
-                                                  mean=overparam_mean, std=overparam_std)
+        self.criterion = overparametrization_loss(
+            num_examp=self.num_training_samples,
+            num_classes=self.num_classes,
+            ratio_consistency=self.ratio_consistency,
+            ratio_balance=ratio_balance,
+            mean=overparam_mean,
[…]
