+++ .pre-commit-config.yaml
@@ -1,6 +1,6 @@
 ﻿repos:
   - repo: https://github.com/pre-commit/pre-commit-hooks
-    rev: v4.6.0
+    rev: v5.0.0
     hooks:
       - id: check-yaml
       - id: check-toml
@@ -10,7 +10,7 @@
       - id: check-merge-conflict
 
   - repo: https://github.com/astral-sh/ruff-pre-commit
-    rev: v0.4.1
+    rev: v0.11.0
     hooks:
       - id: ruff
         args: ["--fix", "--verbose"]
