### 2025-04-04 89a673e
added DISC learning strategy
