+++ src/noisypy/methods/learning_strategies/DISC/DISC.py
@@ -1,0 +1,197 @@
+import torch
+from torch import nn
+from torch.nn import functional as F
+import torchmetrics
+from tqdm import tqdm
+
+from ..base import LearningStrategyModule
+from .utils import GCELoss, Mixup
+
+
+class DISC(LearningStrategyModule):
+
+    def __init__(
+        self,
+        datamodule, classifier_cls, classifier_args, optimizer_cls, optimizer_args, scheduler_cls, scheduler_args,
+        start_epoch: int,
+        alpha: float, # 
+        sigma: float, # 0.5
+        momentum: float, # 0.95
+        lambd_ce: float, # 1
+        lambd_h: float, # 1
+        *args, **kwargs
+    ):
+        super().__init__(datamodule, classifier_cls, classifier_args, optimizer_cls, optimizer_args, scheduler_cls, scheduler_args, *args, **kwargs)
+
+        self.model = classifier_cls(**classifier_args)
+
+        self.start_epoch = start_epoch
+        self.alpha = alpha
+        self.sigma = sigma
+        self.momentum = momentum
+        self.lamd_ce = lambd_ce
+        self.lamd_h = lambd_h
+
+        N = datamodule.num_train_samples
+        C = datamodule.num_classes
+
+        # Variable definition
+        self.s_prev_confidence = torch.ones(N).to(self.device)*1/N
+        self.w_prev_confidence = torch.ones(N).to(self.device)*1/N
+        self.ws_prev_confidence = torch.ones(N).to(self.device)*1/N
+
+        self.w_probs = torch.zeros(N, C).to(self.device)
+        self.s_probs = torch.zeros(N, C).to(self.device)
+        self.labels = torch.ones(N).long().to(self.device)
+        self.weak_labels = self.labels.detach().clone()
+
+        self.clean_flags = torch.zeros(N).bool().to(self.device)
+        self.hard_flags = torch.zeros(N).bool().to(self.device)
+        self.correction_flags = torch.zeros(N).bool().to(self.device)
+        self.weak_flags = torch.zeros(N).bool().to(self.device)
+        self.w_selected_flags = torch.zeros(N).bool().to(self.device)
+        self.s_selected_flags = torch.zeros(N).bool().to(self.device)
+        self.selected_flags = torch.zeros(N).bool().to(self.device)
+        self.class_weight = torch.ones(C).to(self.device)
+        
+        # Loss function definition
+        self.GCE_loss = GCELoss(num_classes=C, gpu="0")
+        self.mixup_loss = Mixup(gpu="0", num_classes=C, alpha=alpha)
+        self.criterion = nn.CrossEntropyLoss()
+
+        # metrics
+        self.train_acc = torchmetrics.Accuracy(num_classes=C, top_k=1, task='multiclass', average="micro")
+        self.val_acc = torchmetrics.Accuracy(num_classes=C, top_k=1, task='multiclass', average="micro")
+        self.test_acc = torchmetrics.Accuracy(num_classes=C, top_k=1, task='multiclass', average="micro")
+
+    def on_train_start(self):
+        self.get_labels(self.datamodule.train_dataloader()[0])
+
+    def training_step(self, batch, batch_idx):
+        w_images, s_images, targets, indexes = batch[0]
+        device = w_images.device
+
+        w_imgs = torch.tensor(w_images).to(self.device)
+        s_imgs = torch.tensor(s_images).to(self.device)
+        targets = torch.tensor(targets).to(self.device)
+
+        all_inputs = torch.cat([w_imgs, s_imgs], dim=0)
+        bs = w_imgs.shape[0]
+        logits = self.model(all_inputs)
+        w_logits = logits[:bs]
+        s_logits = logits[bs:]
+
+
+        if self.current_epoch <= self.start_epoch:
+
+            # ----- warmup ----- #
[…]
