+++ src/reproducibility/noisy_labels/configs/base/wrappers.py
@@ -13,6 +13,10 @@ from noisypy.data.datasets.cifar10 import (
     cifar10_test_transform,
 )
 from noisypy.methods.learning_strategies.SOPplus.utils import autoaug_paper_cifar10
+from noisypy.methods.learning_strategies.DISC.utils import (
+    weak_train_transform,
+    strong_train_transform,
+)
 
 
 def add_index_wrapper(data_config: DataConfig) -> DataConfig:
@@ -54,6 +58,21 @@ def double_aug_wrapper(data_config: DataConfig) -> DataConfig:
                 AddIndex(),
                 DoubleAugmentation(
                     transform1=cifar10_train_transform, transform2=autoaug_paper_cifar10
+                ),
+            ]
+        )
+
+    return new_data_config
+
+
+def disc_aug_wrapper(data_config):
+    class new_data_config(data_config):
+        dataset_train_args = {**data_config.dataset_train_args, "transform": None}
+        dataset_train_augmentation = Compose(
+            [
+                AddIndex(),
+                DoubleAugmentation(
+                    transform1=weak_train_transform, transform2=strong_train_transform
                 ),
             ]
         )

[…]
