### 2025-04-04 f47db7e
fixed DISC config for cifar100n
