+++ src/reproducibility/noisy_labels/configs/benchmark_configs/cifar10n/DISC.py
@@ -7,13 +7,12 @@ from ...base.wrappers import disc_aug_wrapper
 
 
 class DISC_config(BenchmarkConfigCIFAR10N):
-
     _data_config_wrapper = disc_aug_wrapper
 
     learning_strategy_cls = DISC
     learning_strategy_args = dict(
-        start_epoch=15, # https://github.com/JackYFL/DISC/blob/a21e910bffeb34873684937ac1066991a720552a/algorithms/DISC.py#L81
-        alpha=5.0, 
+        start_epoch=15,  # https://github.com/JackYFL/DISC/blob/a21e910bffeb34873684937ac1066991a720552a/algorithms/DISC.py#L81
+        alpha=5.0,
         sigma=0.5,
         momentum=0.99,
         lambd_ce=1.0,
@@ -24,7 +23,7 @@ class DISC_config(BenchmarkConfigCIFAR10N):
     optimizer_args = dict(lr=0.1, momentum=0, weight_decay=0.001)
     scheduler_cls = MultiStepLR
     scheduler_args = dict(milestones=[80, 160])
-    
+
     trainer_args = {
         **BenchmarkConfigCIFAR10N.trainer_args,
         "reload_dataloaders_every_n_epochs": 1,

[…]
