+++ .pre-commit-config.yaml
@@ -10,7 +10,7 @@
       - id: check-merge-conflict
 
   - repo: https://github.com/astral-sh/ruff-pre-commit
-    rev: v0.11.0
+    rev: v0.11.4
     hooks:
       - id: ruff
         args: ["--fix", "--verbose"]
