### 2025-04-09 2adc16e
add ProMix benchmark config
