### 2025-04-09 eaf45bc
update memorization experiment
