+++ src/reproducibility/memorization_effects/memo.py
@@ -1,3 +1,4 @@
+import os
 import argparse
 
 import torch
@@ -6,10 +7,10 @@ from torch.nn import functional as F
 from torchvision.datasets import CIFAR10
 from torchvision import transforms
 from tqdm import tqdm
+import lightning as L
 
-from utils import ResNet34 as resnet34
+from noisypy.methods.classifiers.resnet import ResNet34 as resnet34
 
-device = "cuda" if torch.cuda.is_available() else "cpu"
 
 argparser = argparse.ArgumentParser()
 argparser.add_argument(
@@ -18,16 +19,35 @@ argparser.add_argument(
     help="noise label type",
     choices=["aggre_label", "random_label1", "worse_label"],
 )
+argparser.add_argument(
+    "--scheduler", default="exponential", choices=["exponential", "step"]
+)
+argparser.add_argument("--weight-decay", type=float, default=1e-4, help="weight decay")
 argparser.add_argument("--seed", type=int, default=1, help="random seed")
 argparser.add_argument("--device", type=int, default=0, help="device")
+argparser.add_argument(
+    "--initial-lr", type=float, default=1, help="initial learning rate"
+)
 
 args = argparser.parse_args()
 seed = args.seed
 noise_label = args.noise_label
+
+L.seed_everything(args.seed)
+
+save_dir = os.path.join(os.path.dirname(os.path.abspath(__file__)), args.scheduler)
+if args.weight_decay != 1e-4:
+    save_dir += f"_wd_{args.weight_decay}"
+if args.initial_lr != 1:
+    save_dir += f"_lr_{args.initial_lr}"
+os.makedirs(save_dir, exist_ok=True)
+
 device = f"cuda:{args.device}" if torch.cuda.is_available() else "cpu"
 
 noisy_path = "../../../data/noisylabels/CIFAR-10_human.pt"
-noisy_labels = {k: torch.tensor(v) for k, v in torch.load(noisy_path).items()}
+noisy_labels = {
+    k: torch.tensor(v) for k, v in torch.load(noisy_path, weights_only=False).items()
+}
 
 
 cifar10_train_transform = transforms.Compose(
@@ -43,7 +63,6 @@ clean_dataset = CIFAR10(
     "../../../data/cifar", train=True, transform=cifar10_train_transform
 )
 clean_labels = torch.tensor(clean_dataset.targets)
-clean_labels
 
 
 def generate_synthetic_labels(clean_labels, noisy_labels, seed=1337):
@@ -62,6 +81,46 @@ def generate_synthetic_labels(clean_labels, noisy_labels, seed=1337):
     return synthetic_labels
 
 
+def generate_pairflip_labels(clean_labels, noisy_labels, seed=1337):
+    noise_rate = (clean_labels != human_labels).float().mean()
+    # generate transition matrix T
+    num_classes = 10
+    T = torch.zeros(num_classes, num_classes)
+    T[torch.eye(num_classes, dtype=torch.bool)] = 1 - noise_rate
+    for i in range(num_classes):
+        clean_mask = clean_labels == i
+        # Estimate the original transition vector
+        t = torch.zeros(num_classes)
+        for j in range(num_classes):
+            if i == j:
+                continue
+            p = (noisy_labels[clean_mask] == j).float().mean()
+            t[j] = p
+        # Select the most likely transition class
+        j = t.argmax()
+        T[i, j] = noise_rate
+
+    assert torch.isclose(T.sum(axis=1), torch.ones(num_classes)).all()
+    # sample the transition matrix
+    gen = torch.Generator("cpu").manual_seed(seed)
+    synthetic_labels = torch.multinomial(T[clean_labels], 1, generator=gen).squeeze()
[…]
