+++ README.md
@@ -26,12 +26,12 @@ We implement and test the following experiments:
 | Co-teaching  | $92.15 \pm 0.11$              | $91.93 \pm 0.25$              | $90.69 \pm 0.07$              | $90.51 \pm 0.14$              | $90.56 \pm 0.22$              | $80.92 \pm 0.43$              | $72.24 \pm 0.44$              | $54.48 \pm 0.27$              |
 | Co-teaching+ | $92.90 \pm 0.22$              | $91.15 \pm 0.08$              | $89.82 \pm 0.13$              | $89.64 \pm 0.19$              | $89.77 \pm 0.26$              | $82.36 \pm 0.04$              | $70.39 \pm 0.45$              | $55.46 \pm 0.34$              |
 | ELR          | $93.97 \pm 0.12$              | $93.00 \pm 0.19$              | $92.20 \pm 0.10$              | $92.05 \pm 0.12$              | $92.18 \pm 0.17$              | $87.89 \pm 0.14$              | $75.64 \pm 0.21$              | $63.72 \pm 0.38$              |
-| ELR+         | $95.81 \pm 0.16$              | $95.32 \pm 0.06$              | $94.89 \pm 0.11$              | $94.88 \pm 0.08$              | $94.93 \pm 0.07$              | $91.75 \pm 0.06$              | $\boldsymbol{78.82 \pm 0.24}$ | $67.87 \pm 0.07$              |
-| DivideMix    | $95.51 \pm 0.00$              | $95.62 \pm 0.09$              | $\boldsymbol{95.72 \pm 0.11}$ | $\boldsymbol{95.78 \pm 0.10}$ | $95.71 \pm 0.09$              | $\boldsymbol{93.10 \pm 0.10}$ | $78.22 \pm 0.06$              | $\boldsymbol{70.91 \pm 0.09}$ |
+| ELR+         | $95.81 \pm 0.16$              | $95.32 \pm 0.06$              | $94.89 \pm 0.11$              | $94.88 \pm 0.08$              | $94.93 \pm 0.07$              | $91.75 \pm 0.06$              | $\mathbf{78.82 \pm 0.24}$ | $67.87 \pm 0.07$              |
+| DivideMix    | $95.51 \pm 0.00$              | $95.62 \pm 0.09$              | $\mathbf{95.72 \pm 0.11}$ | $\mathbf{95.78 \pm 0.10}$ | $95.71 \pm 0.09$              | $\mathbf{93.10 \pm 0.10}$ | $78.22 \pm 0.06$              | $\mathbf{70.91 \pm 0.09}$ |
 | VolMinNet    | $92.71 \pm 0.02$              | $90.47 \pm 0.17$              | $88.90 \pm 0.51$              | $88.81 \pm 0.17$              | $88.67 \pm 0.10$              | $80.87 \pm 0.25$              | $72.73 \pm 0.65$              | $58.30 \pm 0.05$              |
 | CAL          | $93.78 \pm 0.18$              | $91.84 \pm 0.23$              | $91.10 \pm 0.26$              | $90.60 \pm 0.10$              | $90.61 \pm 0.12$              | $84.82 \pm 0.23$              | $74.53 \pm 0.21$              | $60.13 \pm 0.33$              |
 | PES (semi)   | $94.75 \pm 0.16$              | $94.64 \pm 0.05$              | $95.20 \pm 0.08$              | $95.26 \pm 0.13$              | $95.20 \pm 0.11$              | $92.58 \pm 0.05$              | $77.77 \pm 0.33$              | $70.32 \pm 0.28$              |
[…]
