### 2025-10-05 e8c2b88
Update and refactor (#16)
* update benchmark scripts

* update DivideMix data wrapper

* Implement ProMix

* Lint and add pre-commit hook

* autoupdate pre-commit hooks

* add cutout to ProMix

* Add DISC LNL (#15)

Co-authored-by: Ziga R <102856773+rziga@users.noreply.github.com>

* bump ruff version

* refactor

* add ProMix benchmark config

* update memorization experiment

* update memorization results

* upload DISC and ProMix logs

* format

* update memorization figure

---------

Co-authored-by: Ziga R <102856773+rziga@users.noreply.github.com>
