+++ src/reproducibility/learning_strategies/cores2/cores_cifar10_clean.py
@@ -13,27 +13,22 @@ from ..common import cifar10_base_config
 
 lr_plan = [0.1] * 50 + [0.01] * (50 + 1)
 
-class cores_cifar10_clean(MethodConfig):
 
+class cores_cifar10_clean(MethodConfig):
     data_config = cifar10_base_config
 
     classifier = resnet34
-    classifier_args = dict(
-        weights=None,
-        num_classes=10
-    )
+    classifier_args = dict(weights=None, num_classes=10)
 
     learning_strategy_cls = SampleSieve
 
     optimizer_cls = SGD
     optimizer_args = dict(lr=0.1, momentum=0.9, weight_decay=5e-4)
     scheduler_cls = LambdaLR
-    scheduler_args = dict(lr_lambda=lambda epoch: lr_plan[epoch] / (1+f_beta(epoch)))
+    scheduler_args = dict(lr_lambda=lambda epoch: lr_plan[epoch] / (1 + f_beta(epoch)))
 
     trainer_args = dict(
-        max_epochs=100,
-        deterministic=True,
-        logger=CSVLogger("../logs", name="cores2")
+        max_epochs=100, deterministic=True, logger=CSVLogger("../logs", name="cores2")
     )
 
     seed = 1337
