### 2025-10-15 5c2325d
prepare for ReScience submission (#17)
* add metadata for rescience submission

* upload pdf
