### 2025-10-20 #86 @ValterH
> Ok, are there any public reviews by any chance?

Yes, the reviews are available at https://openreview.net/forum?id=GKZ2leags0
