### 2025-10-20 #86 @rougier
Thanks. Based on these public reviews, I think we can accept directly your paper for publication.

 My only two comments concerns the title: [Re] means full replication, [~Re] means partial replication and [¬Re] means not reproducible. Based on your discussion / conclusion, do you prefere to keep [Re] or to use [~Re] ? the other very minor comment is in the text with a lot line returns. Is that intended ?
