### 2025-10-20 #86 @ValterH
Thank you for the quick response!

> My only two comments concerns the title: [Re] means full replication, [~Re] means partial replication and [¬Re] means not reproducible. Based on your discussion / conclusion, do you prefere to keep [Re] or to use [~Re] ?

I agree that a partial reproduction is more approporiate, hence we will change the name to use  [~Re].

> the other very minor comment is in the text with a lot line returns. Is that intended ?

I added breaks between paragraphs to replace the line returns.

Both changes are reflected in the updated version of the article: https://github.com/KlemenVovk/noisy-labels/blob/rescience/rescience/rescience.pdf
