### 2026-01-08 #86 @rougier
Sorry for long delay, end of year was a bit crazy. I started the publication but then I realized the latex sources are not present in your repo. Can you add them (with the makefile / cls / etc). I'll need to edit it to add the DOI and update volume issue.
(latest template is here: https://github.com/ReScience/template)

In the meantime, can you also save your github repo on software heritage?
