+++ rescience/tables/benchmark.tex
@@ -1,0 +1,24 @@
+\begin{tabular}{ccccccc|cc}
+\toprule
+ \multirow{2}{*}{\textbf{Method}} & \multicolumn{6}{c|}{\textbf{CIFAR-10N}} & \multicolumn{2}{c}{\textbf{CIFAR-100N}} \\
+ & Clean & Aggregate & Random 1 & Random 2 & Random 3 & Worst & \multicolumn{1}{c}{Clean} & \multicolumn{1}{c}{Noisy} \\
+\midrule
+
+CE & $85.50$\footnotesize{$  \pm 0.25 $} & $83.50$\footnotesize{$  \pm 0.23 $} & $81.96$\footnotesize{$  \pm 0.77 $} & $82.00$\footnotesize{$  \pm 0.13 $} & $82.19$\footnotesize{$  \pm 0.31 $} & $75.48$\footnotesize{$  \pm 0.30 $} & $59.18$\footnotesize{$  \pm 0.18 $} & $49.51$\footnotesize{$  \pm 0.39 $} \\
+ELR+ & $88.56$\footnotesize{$  \pm 0.23 $} & $87.76$\footnotesize{$  \pm 0.16 $} & $86.75$\footnotesize{$  \pm 0.15 $} & $86.94$\footnotesize{$  \pm 0.07 $} & $86.97$\footnotesize{$  \pm 0.07 $} & $81.50$\footnotesize{$  \pm 0.12 $} & $61.80$\footnotesize{$  \pm 0.25 $} & $52.91$\footnotesize{$  \pm 0.27 $} \\
+VolMinNet & $82.57$\footnotesize{$  \pm 0.20 $} & $80.42$\footnotesize{$  \pm 0.36 $} & $79.18$\footnotesize{$  \pm 0.10 $} & $78.54$\footnotesize{$  \pm 0.30 $} & $78.56$\footnotesize{$  \pm 0.13 $} & $72.07$\footnotesize{$  \pm 0.49 $} & $51.27$\footnotesize{$  \pm 0.18 $} & $42.39$\footnotesize{$  \pm 0.74 $} \\
+CAL & $83.38$\footnotesize{$  \pm 0.47 $} & $81.51$\footnotesize{$  \pm 0.21 $} & $79.64$\footnotesize{$  \pm 0.45 $} & $79.50$\footnotesize{$  \pm 0.31 $} & $79.67$\footnotesize{$  \pm 0.26 $} & $73.08$\footnotesize{$  \pm 0.88 $} & $58.02$\footnotesize{$  \pm 0.17 $} & $47.37$\footnotesize{$  \pm 0.49 $} \\
+PES (semi) & $87.08$\footnotesize{$  \pm 0.08 $} & $86.62$\footnotesize{$  \pm 0.23 $} & $87.63$\footnotesize{$  \pm 0.11 $} & $87.47$\footnotesize{$  \pm 0.20 $} & $86.80$\footnotesize{$  \pm 0.66 $} & $84.12$\footnotesize{$  \pm 0.34 $} & $60.51$\footnotesize{$  \pm 0.14 $} & $52.95$\footnotesize{$  \pm 0.47 $} \\
+SOP & $82.12$\footnotesize{$  \pm 0.43 $} & $79.78$\footnotesize{$  \pm 0.85 $} & $79.22$\footnotesize{$  \pm 0.57 $} & $79.37$\footnotesize{$  \pm 0.29 $} & $78.66$\footnotesize{$  \pm 0.57 $} & $71.76$\footnotesize{$  \pm 1.98 $} & $52.61$\footnotesize{$  \pm 0.43 $} & $40.98$\footnotesize{$  \pm 0.28 $} \\
[…]
