+++ rescience/tables/comparison.tex
@@ -1,0 +1,16 @@
+\begin{tabular}{ccc|c}
+    \toprule
+    \textbf{Method} & \textbf{Described Config} & \textbf{Original Config} & \textbf{Reported} \\
+    \midrule
+    CE              & $91.70$\tiny{$  \pm 0.07 $}  & $91.70$\tiny{$  \pm 0.07 $}  & $87.77$\tiny{$  \pm 0.38 $} \\
+    Co-teaching     & $90.25$\tiny{$  \pm 0.13 $}  & $91.93$\tiny{$  \pm 0.25 $}  & $91.20$\tiny{$  \pm 0.13 $} \\
+    Co-teaching+    & $86.20$\tiny{$  \pm 0.88 $}  & $91.15$\tiny{$  \pm 0.08 $}  & $90.61$\tiny{$  \pm 0.22 $} \\
+    ELR$^*$             & $93.00$\tiny{$  \pm 0.19 $}  & $93.00$\tiny{$  \pm 0.19 $}  & $92.38$\tiny{$  \pm 0.64 $} \\
+    ELR+$^*$            & $95.32$\tiny{$  \pm 0.06 $}  & $95.32$\tiny{$  \pm 0.06 $}  & $94.83$\tiny{$  \pm 0.10 $} \\
+    DivideMix$^*$       & $95.62$\tiny{$  \pm 0.09 $}  & $95.62$\tiny{$  \pm 0.09 $}  & $95.01$\tiny{$  \pm 0.71 $} \\
+    VolMinNet       & $84.02$\tiny{$  \pm 6.61 $}  & $90.47$\tiny{$  \pm 0.17 $}  & $89.70$\tiny{$  \pm 0.21 $} \\
+    CAL             & $91.76$\tiny{$  \pm 0.22 $}  & $91.84$\tiny{$  \pm 0.23 $}  & $91.97$\tiny{$  \pm 0.32 $} \\
+    PES (semi)      & $91.53$\tiny{$  \pm 0.44 $}  & $94.64$\tiny{$  \pm 0.05 $}  & $94.66$\tiny{$  \pm 0.18 $} \\
+    SOP+ & $93.17$\tiny{$  \pm 0.66 $}  & $96.04$\tiny{$  \pm 0.15 $}  & $95.61$\tiny{$  \pm 0.13 $} \\
+    \bottomrule
+    \end{tabular}
