+++ rescience/appendix.tex
@@ -1,0 +1,298 @@
+\section{Baseline Performance Results} \label{appsec:ce}
+
+The largest discrepancy between our and the authors' results comes from the baseline (CE) training. \cite{noisylabels-benchmark}, do not describe the exact procedure for obtaining their results. In our correspondence (see Appendix~\ref{appsec:communication}), they clarify that they used the best test set performance and report the mean and standard deviation across five random seeds. This procedure results in a discrepancy between our reproduced results and the ones reported in the original work~\citep{noisylabels-benchmark}.
+We visualize the discrepancy in Figure \ref{fig:CE-difference}.  
+\newline
+
+We can see that the accuracy of the best checkpoint in our experiments is higher than their reported performance across all runs. We were not alone in noticing this discrepancy, as at the time of investigation, there were two open issues on the authors' official Github repository with the same question. Authors of the issues report the last obtained accuracies on the \textit{Worst} label set:$ \approx 68 \%$ \footnote{\url{https://github.com/UCSC-REAL/cifar-10-100n/issues/5}} and $ \approx 67 \%$ \footnote{\url{https://github.com/UCSC-REAL/cifar-10-100n/issues/8}}, which are also in line with our last epoch accuracy ($68 \pm 0.88 $).
+
+\begin{figure}[!ht]
+    \centering
+    \includegraphics[width=0.8\textwidth]{figures/CE_comparison.pdf}
+    \caption{\textbf{Differences between the authors' reported (dotted) baseline performance and accuracy obtained in reproduction (full blue) as well as using the original code (orange, red, green).} The performance obtained using a reimplemented framework perfectly aligns with the original code. However, on all the noise labels, their reported accuracy does not coincide with the best accuracy for any of the runs.}
+    \label{fig:CE-difference}
+\end{figure}
+
+
+\section{Argument For Using Accuracy On Noisy Data} \label{appsec:accuracy}
+
+% Why accuracy on clean test is not good
+Since testing every epoch and reporting the highest result overestimates performance and is thus considered a bad practice, we have to find an alternative. We want to use the established train-validation-test split in this case. The first problem in LNL setting arises when we need to decide whether the validation set is noisy or not, since practices for this are not yet established in the field. Given the nature of the task, we argue that the validation set should be noisy as well~\citep{jiang2020beyond}. If one has non-noisy data, one can train on it, implicitly changing the noise ratio of the data. Additionally, some LNL methods \citep{forwardbackwardt, f-divergence} use validation data to estimate parameters for loss functions that are later used to train. If validation data were to be clean, there would be a direct data leak in such cases.
+\newline
+
+% Why we pick accuracy on noisy val
+Consequently, we must decide on a proxy metric for the accuracy on clean data. We consider two options: using validation losses of each LNL method or validation accuracy. Given that losses vary between the LNL methods (some have no upper bounds) and are not as stable (see Figure~\ref{'fig:val-acc-loss'}), we propose to use accuracy on noisy validation data for model selection. We propose this, as it remains stable across all tested methods and strongly correlates with final test performance, preserving model ranking with high probability.
+\newline
+
+\begin{figure}[!ht]
+    \centering
[…]
