+++ rescience/article.tex
@@ -1,0 +1,22 @@
+\documentclass{rescience}
+\usepackage{booktabs}
+\usepackage{amsfonts}
+\usepackage{subcaption}
+\usepackage{multirow}
+\usepackage{longtable}
+\usepackage{lipsum}
+\usepackage[]{mdframed}
+
+\begin{document}
+\input{header.tex}
+\input{content.tex}
+
+\hypersetup{linkcolor=black,urlcolor=darkgray}
+\renewcommand\emph[1]{{\bfseries #1}}
+\setlength\bibitemsep{0pt}
+\printbibliography
+
+\appendix
+\input{appendix.tex}
+
+\end{document}
