+++ .pre-commit-config.yaml
@@ -15,3 +15,4 @@
       - id: ruff
         args: ["--fix", "--verbose"]
       - id: ruff-format
+exclude: 'rescience/*'

[…]
