### 2026-01-08 e669b50
upload latex content (#19)
