+++ rescience/article.py
@@ -3,7 +3,6 @@
 
 import yaml
 
-
 class Contributor:
     def __init__(self, role, name, orcid="", email="", affiliations=[]):
         self.role = role
@@ -16,10 +15,9 @@ class Contributor:
         self.affiliations = affiliations
 
     def get_abbrvname(self, name):
-        if not name:
-            return ""
-
-        if "," in name:
+        if not name: return ""
+        
+        if ',' in name:
             lastname = name.split(",")[0]
             firstnames = name.split(",")[1].strip().split(" ")
         else:
@@ -29,32 +27,31 @@ class Contributor:
         for firstname in firstnames:
             if "-" in firstname:
                 for name in firstname.split("-"):
-                    abbrvname += name[0].strip().upper() + ".-"
+                    abbrvname += name[0].strip().upper() + '.-'
                 abbrvname = abbrvname[:-1]
             else:
-                abbrvname += firstname[0].strip().upper() + "."
+                abbrvname += firstname[0].strip().upper() + '.'
         return abbrvname + " " + lastname
 
+
     def get_lastname(self, name):
-        if not name:
-            return ""
+        if not name: return ""
         # Rougier, Nicolas P.
-        if "," in name:
+        if ',' in name:
             lastname = name.split(",")[0].strip()
-            # firstname = name.split(",")[1].strip()
+            firstname = name.split(",")[1].strip()
         # Nicolas P. Rougier
         else:
             lastname = name.split(" ")[-1]
-            # firstname = name.split(" ")[:-1]
+            firstname = name.split(" ")[:-1]
         return lastname
-
+    
 
 class Affiliation:
     def __init__(self, code, name, address=""):
         self.code = code
         self.name = name
         self.address = address
-
 
 class Repository:
     def __init__(self, name, url, doi, swh=""):
@@ -63,7 +60,6 @@ class Repository:
         self.doi = doi
         self.swh = swh
 
-
 class Replication:
     def __init__(self, cite, bib, url, doi):
         self.cite = cite
@@ -71,12 +67,10 @@ class Replication:
         self.url = url
         self.doi = doi
 
-
 class Review:
     def __init__(self, url, doi):
         self.url = url
         self.doi = doi
-
 
 class Date:
     def __init__(self, date):
@@ -89,9 +83,8 @@ class Date:
             self.month = date.month
             self.day = date.day
             self.textual = self.date.strftime("%d %B %Y")
-        except dateutil.parser._parser.ParserError:
+        except:
             import datetime
-
             now = datetime.datetime.now()
             self.date = now
             self.year = now.year
@@ -101,12 +94,12 @@ class Date:
 
     def __str__(self):
         return self.textual
-        # return self.date.strftime("%d %B %Y")
+        #return self.date.strftime("%d %B %Y")
 
     def __repr__(self):
         return self.textual
         # return self.date.strftime("%d %B %Y")
-
+        
 
 class Article:
     def __init__(self, data):
@@ -127,12 +120,12 @@ class Article:
 
         self.review = ""
         self.replication = ""
-
+        
         self.date_received = ""
         self.date_accepted = ""
         self.date_published = ""
 
-        self.journal_name = ""
+        self.journal_name = "" 
         self.journal_issn = ""
         self.journal_volume = ""
         self.journal_issue = ""
@@ -143,40 +136,42 @@ class Article:
         self.parse(data)
 
         # Build authors list
-        self.authors_short = ""  # Family names only
-        self.authors_abbrv = ""  # Abbreviated firsnames + Family names
+        self.authors_short = "" # Family names only
+        self.authors_abbrv = "" # Abbreviated firsnames + Family names
         self.authors_full = ""  # Full names
-
+        
[…]
