### 2026-01-08 #86 @ValterH
> Sorry for long delay, end of year was a bit crazy.

We understand. Happy New Year!

> I started the publication but then I realized the latex sources are not present in your repo. Can you add them (with the makefile / cls / etc). I'll need to edit it to add the DOI and update volume issue. (latest template is here: https://github.com/ReScience/template)

Done all the source files are in the [rescience directory ](https://github.com/KlemenVovk/noisy-labels/tree/master/rescience).

> In the meantime, can you also save your github repo on software heritage?

We submitted [a request](https://archive.softwareheritage.org/browse/origin/directory/?origin_url=https://github.com/KlemenVovk/noisy-labels&snapshot=8f9eb576d5247d04ec04cd3c143d3c673f571397). Please let me know if you need some further action on our end.
