### 2026-01-09 #86 @rougier
I just uploaded a **sandbox** version at: https://sandbox.zenodo.org/record/423720. Can you check everything is ok? (authors, links, etc). Once you get the swhid I can publish the final version.
