### 2026-01-09 #86 @ValterH
Thank you for the quick response.

We found that everything seems to be ok. For the title, however, based on our previous discussion, we agree that a partial reproduction [~Re] would be more appropriate. I see that the pdf I uploaded failed to reflect this change. The updated version now does: https://github.com/KlemenVovk/noisy-labels/blob/master/rescience/rescience.pdf

The SWHID is: `swh:1:dir:22222955c66aeb7e379a16b45d208fcf3d64a11d` at [url](https://archive.softwareheritage.org/browse/origin/directory/?origin_url=https://github.com/KlemenVovk/noisy-labels).
