### 2026-01-28 #86 @rougier
Sorry for the delay. Article is online at https://zenodo.org/record/18401497 and shoudl appear  soon on ReScience website.
